services/training/requirements.txt
clearml>=1.2.0
psycopg2-binary>=2.9.9
minio>=7.1.0
scikit-learn>=1.3.0
numpy>=1.24.0